+++ tests/MessengerQueryBusTest.php
@@ -5,12 +5,13 @@ declare(strict_types=1);
 namespace Tuzex\Cqrs\Test;
 
 use PHPUnit\Framework\TestCase;
+use stdClass;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Stamp\HandledStamp;
 use Tuzex\Cqrs\Exception\QueryHandlerNotFoundException;
-use Tuzex\Cqrs\Exception\QueryResultNotFoundException;
+use Tuzex\Cqrs\Exception\InvalidQueryResultException;
 use Tuzex\Cqrs\MessengerQueryBus;
 use Tuzex\Cqrs\Query;
 use Tuzex\Cqrs\QueryHandler;
@@ -20,32 +21,50 @@ final class MessengerQueryBusTest extends TestCase
     public function testItReturnsQueryResult(): void
     {
         $query = $this->createMock(Query::class);
-        $queryBus = new MessengerQueryBus($this->mockMessageBus($query));
+        $queryBus = new MessengerQueryBus(
+            $this->mockMessageBus($query, [$this->mockHandledStamp()])
+        );
 
-        $this->assertInstanceOf(Query::class, $queryBus->execute($query));
+        $this->assertInstanceOf(stdClass::class, $queryBus->execute($query));
     }
 
     public function testItThrowsExceptionIfQueryHandlerNotExists(): void
     {
         $query = $this->createMock(Query::class);
-        $queryBus = new MessengerQueryBus($this->mockMessageBus($query, false));
+        $queryBus = new MessengerQueryBus($this->mockMessageBus($query, handle: false));
 
         $this->expectException(QueryHandlerNotFoundException::class);
         $queryBus->execute($query);
     }
 
-    public function testItThrowsExceptionIfQueryResultMissing(): void
+    /**
+     * @dataProvider provideInvalidStamps
+     */
+    public function testItThrowsExceptionIfQueryResultIsInvalid(array $stamps): void
     {
         $query = $this->createMock(Query::class);
-        $queryBus = new MessengerQueryBus($this->mockMessageBus($query, true, false));
+        $queryBus = new MessengerQueryBus($this->mockMessageBus($query, $stamps));
 
-        $this->expectException(QueryResultNotFoundException::class);
+        $this->expectException(InvalidQueryResultException::class);
         $queryBus->execute($query);
     }
 
-    private function mockMessageBus(Query $query, bool $handle = true, bool $result = true): MessageBusInterface
+    public function provideInvalidStamps(): array
     {
-        $stamps = $result ? [new HandledStamp($query, QueryHandler::class)] : [];
+        return [
+            'not-exists' => [
+                'stamps' => [],
+            ],
+            'invalid-type' => [
+                'stamps' => [
+                    $this->mockHandledStamp(valid: false),
+                ],
+            ],
+        ];
+    }
+
+    private function mockMessageBus(Query $query, array $stamps = [], bool $handle = true): MessageBusInterface
+    {
         $envelope = new Envelope($query, $stamps);
 
         $messageBus = $this->createMock(MessageBusInterface::class);
@@ -58,5 +77,10 @@ final class MessengerQueryBusTest extends TestCase
         }
 
         return $messageBus;
+    }
+
+    private function mockHandledStamp(bool $valid = true): HandledStamp
+    {
+        return new HandledStamp($valid ? new stdClass() : '', QueryHandler::class);
     }
 }