+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
-- **[BC BREAK]** Bump minimal PHP version to 8.*
+- **[BC BREAK]** Bump minimal PHP to version 8.*
 - Use constructor property promotion
+- Check the return type in  ```MessengerQueryBus```
 - Remove obsolete ```get_class``` from exceptions
 
 [Unreleased]: https://github.com/Tuzex/responder/compare/v1.0.0...HEAD